// ui/src/lib/types/api.types.ts
export interface components {
    schemas: {
// ...
        ProductConstraintsDto: {
            minAmount?: number;
            maxAmount?: number;
            minTerm?: number;
            maxTerm?: number;
            minRate?: number;
            maxRate?: number;
        };
        ProductDefaultsDto: {
            fixedRate?: number;
            /** @default USD */
            currency: components["schemas"]["CurrencyCode"];
            /** @default 2 */
            roundingScale: number;
            /** @default HALF_UP */
            roundingMode: components["schemas"]["RoundingMode"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}